pyproject.toml
[project]
name = "kalshi-bot"
version = "0.1.0"
description = "Production-ready Discord bot for real-time Kalshi order monitoring and notifications"
requires-python = ">=3.11"
dependencies = [
  "websockets>=12.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "cryptography>=41.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "discord.py>=2.3.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "typer>=0.9.0",
]

[project.scripts]
kalshi-bot = "src.cli:cli"

